+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - fix several 'overflow when subtracting durations' panics
+- [cpu] fix CPUs overflowing widget
 
 ## [0.4.1] - 2020-02-04